extracted layout setup into a separate file
