+++ src/identity/setup.tsx
@@ -1,10 +1,12 @@
 import { setupIdentityUrls } from '@identity/urls.ts';
+import { setupIdentityLayouts } from '@identity/layouts.ts';
 import { setupIdentityMenus } from '@identity/menus.tsx';
 import { setupIdentityServices } from '@identity/services.ts';
 import { setupIdentityViews } from '@identity/views.tsx';
 
 export default function identitySetup() {
 	setupIdentityUrls();
+	setupIdentityLayouts();
 	setupIdentityMenus();
 	setupIdentityServices();
 	setupIdentityViews();
